INTERLOK-3465  - Remove AML from FMR interface
Remove AdaptrisMessageListener from FailedMessageRetrier interface.
Since DefaultFailedMessageRetrier is the only concrete impl in our
source code set; this is "low risk".
