+++ interlok-core/src/main/java/com/adaptris/core/DefaultFailedMessageRetrier.java
@@ -1,12 +1,12 @@
 /*
  * Copyright 2015 Adaptris Ltd.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -24,30 +24,40 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.TimeUnit;
+import java.util.function.Consumer;
 import javax.management.MalformedObjectNameException;
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 import com.adaptris.annotation.AdapterComponent;
 import com.adaptris.annotation.AdvancedConfig;
+import com.adaptris.annotation.AutoPopulated;
 import com.adaptris.annotation.ComponentProfile;
 import com.adaptris.annotation.InputFieldDefault;
 import com.adaptris.core.runtime.AdapterManager;
 import com.adaptris.core.runtime.ParentRuntimeInfoComponent;
 import com.adaptris.core.runtime.RuntimeInfoComponent;
 import com.adaptris.core.runtime.RuntimeInfoComponentFactory;
+import com.adaptris.core.util.Args;
+import com.adaptris.core.util.LifecycleHelper;
+import com.adaptris.core.util.LoggingHelper;
 import com.adaptris.core.util.ManagedThreadFactory;
 import com.adaptris.util.TimeInterval;
 import com.thoughtworks.xstream.annotations.XStreamAlias;
+import lombok.Getter;
+import lombok.Setter;
 
 /**
  * <p>
  * Implementation of <code>FailedMessageRetrier</code> that does not allow duplicate workflows.
  * </p>
- * 
+ *
  * @config default-failed-message-retrier
  */
 @XStreamAlias("default-failed-message-retrier")
 @AdapterComponent
 @ComponentProfile(summary = "A Configurable Failed Message Retrier", tag = "error-handling,base")
-public class DefaultFailedMessageRetrier extends FailedMessageRetrierImp {
+public class DefaultFailedMessageRetrier extends FailedMessageRetrierImp
+    implements AdaptrisMessageListener {
   /**
    * The default shutdown wait.
    *
@@ -59,20 +69,46 @@ public class DefaultFailedMessageRetrier extends FailedMessageRetrierImp {
 
   private transient ExecutorService failedMessageExecutor = null;
 
+  @NotNull
+  @Valid
+  @AutoPopulated
+  @Getter
+  private StandaloneConsumer standaloneConsumer;
+
+  /**
+   * Set the shutdown wait timeout for the pool.
+   * <p>
+   * When {@link #close()} is invoked, this shutdowns the internal retry submission pool used by
+   * {@link DefaultFailedMessageRetrierJmxMBean}. The specified value is the amount of time to wait
+   * for a clean shutdown. If this timeout is exceeded then a forced shutdown ensues, which may mean
+   * some messages will not have been retried. The default is 60 seconds if not explicitly
+   * specified.
+   * </p>
+   */
   @InputFieldDefault(value = "60 seconds")
   @AdvancedConfig(rare = true)
+  @Getter
+  @Setter
   private TimeInterval shutdownWaitTime;
 
-  @Override
-  public void addWorkflow(Workflow workflow) throws CoreException {
-    String key = workflow.obtainWorkflowId();
-    if (getWorkflows().keySet().contains(key)) {
-      log.warn("duplicate workflow ID [" + key + "]");
-      throw new CoreException("Workflows cannot be uniquely identified");
-    }
-    log.debug("adding workflow with key [" + key + "]");
-    getWorkflows().put(key, workflow);
+  public DefaultFailedMessageRetrier() {
+    setStandaloneConsumer(new StandaloneConsumer());
   }
+
+
+  @Override
+  public synchronized void onAdaptrisMessage(AdaptrisMessage msg, Consumer<AdaptrisMessage> success,
+      Consumer<AdaptrisMessage> failure) {
+    try {
+      Workflow workflow = getWorkflow(msg);
+      updateRetryCountMetadata(msg);
+      workflow.onAdaptrisMessage(msg, success, failure); // workflow.onAM is sync'd...
+    } catch (Exception e) { // inc. runtime, exc. Workflow
+      log.error("exception retrying message", e);
+      log.error("message {}", MessageLoggerImpl.LAST_RESORT_LOGGER.toString(msg));
+    }
+  }
+
 
   boolean retryMessage(final AdaptrisMessage msg) {
     boolean submitted = false;
@@ -110,6 +146,57 @@ public class DefaultFailedMessageRetrier extends FailedMessageRetrierImp {
     return retryMessage(msg);
   }
 
+  @Override
+  public void prepare() throws CoreException {
+    LifecycleHelper.prepare(getStandaloneConsumer());
+  }
+
+  @Override
+  public void init() throws CoreException {
+    if (standaloneConsumer != null) {
+      standaloneConsumer.registerAdaptrisMessageListener(this);
+    }
+    LifecycleHelper.init(standaloneConsumer);
+    failedMessageExecutor = Executors.newSingleThreadExecutor();
+  }
+
+  /** @see com.adaptris.core.AdaptrisComponent#start() */
+  @Override
+  public void start() throws CoreException {
+    LifecycleHelper.start(standaloneConsumer);
+  }
+
+  @Override
+  public void stop() {
+    LifecycleHelper.stop(standaloneConsumer);
+  }
+
+  @Override
+  public void close() {
+    LifecycleHelper.close(standaloneConsumer);
+    shutdownExecutors();
+  }
+
+  private void shutdownExecutors() {
+    ManagedThreadFactory.shutdownQuietly(failedMessageExecutor, shutdownWaitTimeMs());
+  }
+
+  private long shutdownWaitTimeMs() {
+    return TimeInterval.toMillisecondsDefaultIfNull(getShutdownWaitTime(), DEFAULT_SHUTDOWN_WAIT);
+  }
+
+  public void setStandaloneConsumer(StandaloneConsumer consumer) {
+    standaloneConsumer = Args.notNull(consumer, "consumer");
+    standaloneConsumer.registerAdaptrisMessageListener(this);
+  }
+
+
+  @Override
+  public String friendlyName() {
+    return LoggingHelper.friendlyName(this);
+  }
+
+
   private static class JmxFactory extends RuntimeInfoComponentFactory {
 
     @Override
@@ -121,59 +208,11 @@ public class DefaultFailedMessageRetrier extends FailedMessageRetrierImp {
     }
 
     @Override
-    protected RuntimeInfoComponent createComponent(ParentRuntimeInfoComponent parent, AdaptrisComponent e)
-        throws MalformedObjectNameException {
-      return new DefaultFailedMessageRetrierJmx((AdapterManager) parent, (DefaultFailedMessageRetrier) e);
+    protected RuntimeInfoComponent createComponent(ParentRuntimeInfoComponent parent,
+        AdaptrisComponent e) throws MalformedObjectNameException {
+      return new DefaultFailedMessageRetrierJmx((AdapterManager) parent,
+          (DefaultFailedMessageRetrier) e);
     }
-  }
-
-  @Override
-  public void init() throws CoreException {
-    super.init();
-    failedMessageExecutor = Executors.newSingleThreadExecutor();
-  }
-
-  @Override
-  public void start() throws CoreException {
-    super.start();
-  }
-
-  @Override
-  public void stop() {
-    super.stop();
-  }
-
-  @Override
-  public void close() {
-    shutdownExecutors();
-    super.close();
-  }
-
-  private void shutdownExecutors() {
-    ManagedThreadFactory.shutdownQuietly(failedMessageExecutor, shutdownWaitTimeMs());
-  }
-
-  private long shutdownWaitTimeMs() {
-    return TimeInterval.toMillisecondsDefaultIfNull(getShutdownWaitTime(), DEFAULT_SHUTDOWN_WAIT);
-  }
-
-  public TimeInterval getShutdownWaitTime() {
-    return shutdownWaitTime;
-  }
-
-  /**
-   * Set the shutdown wait timeout for the pool.
-   * <p>
-   * When {@link #close()} is invoked, this shutdowns the internal retry submission pool used by
-   * {@link DefaultFailedMessageRetrierJmxMBean}. The specified value is the amount of time to wait for a clean shutdown. If this
-   * timeout is exceeded then a forced shutdown ensues, which may mean some messages will not have been retried.
-   * </p>
-   * 
-   * @param interval the shutdown time (default is 60 seconds)
-   * @see #stop()
-   */
-  public void setShutdownWaitTime(TimeInterval interval) {
-    shutdownWaitTime = interval;
   }
 
 }
